fix №20

+++ server/controllers/comments.controller.js
@@ -7,10 +7,9 @@ const commentsController = {
         WITH RECURSIVE cte AS (
           SELECT * FROM unnest($1::json[]) as item
         ), recursive_cte AS (
-          SELECT item ->> 'taskId' as task_id, item ->> 'comment' as comment_text, item ->> 'parentId' as parent_id FROM cte WHERE item ->> 'parentId' IS NULL
+          SELECT (item ->> 'taskId')::integer as task_id, item ->> 'comment' as comment_text, item ->> 'parentId' as parent_id FROM cte WHERE item ->> 'parentId' IS NULL
           UNION ALL
-          SELECT cte.item ->> 'taskId', cte.item ->> 'comment', cte.item ->> 'parentId' FROM cte JOIN recursive_cte ON cte.item ->> 'parentId' = recursive_cte.task_id
-        )
+          SELECT (cte.item ->> 'taskId')::integer, cte.item ->> 'comment', cte.item ->> 'parentId' FROM cte JOIN recursive_cte ON cte.item ->> 'parentId' = recursive_cte.task_id
         INSERT INTO comments (task_id, comment_text, parent_id)
         SELECT task_id, comment_text, parent_id FROM recursive_cte
         RETURNING *;
